requirements.txt
aioinflux==0.9.0
attrs==21.2.0
colorama==0.4.4
frozendict==2.0.2
influxdb==5.3.1
nest-asyncio==1.5.1
pandas==1.2.4
pint==0.17
python-pushover==0.4
pyyaml==5.4.1
ratelimit==2.2.1
regex==2021.4.4
tenacity==7.0.0
tzlocal==2.1
